refactor: simplify request attestation API
