+++ .changeset/attested-agents-pay.md
@@ -2,4 +2,4 @@
 'mppx': patch
 ---
 
-Added composable TAP and Web Bot Auth request-attestation adapters with Ed25519 and RSA-PSS SHA-512 signatures, explicit replay storage, complete verifier outcomes, and shared per-attempt signing context.
+Added composable TAP and Web Bot Auth request signers and verifiers with Ed25519 and RSA-PSS SHA-512 support, explicit replay storage, typed verification outcomes, and shared per-attempt signing context.
