+++ src/attestation/Constants.ts
@@ -1,15 +1,3 @@
-/** Capabilities an attestation protocol can cryptographically establish. */
-export const Capabilities = {
-  /** Identifies the automated client that signed the request. */
-  agentIdentity: 'agent-identity',
-  /** Establishes that the client expressed a protocol-defined commerce intent. */
-  commerceIntent: 'commerce-intent',
-  /** Binds a unique signature nonce to this request. */
-  replayProtection: 'replay-protection',
-  /** Binds signature evidence to selected HTTP request components. */
-  requestBinding: 'request-binding',
-} as const
-
 /** RFC 9421 signature algorithms supported by request attestation. */
 export const Algorithms = {
   /** EdDSA using the Ed25519 curve. */
