+++ src/tap/index.ts
@@ -1,5 +1,4 @@
 export * as Client from './client/index.js'
-export * as Policy from './Policy.js'
 export * as Server from './server/index.js'
 export * from './Constants.js'
 export * from './Types.js'
