+++ src/web-bot-auth/Types.ts
@@ -1,6 +1,3 @@
-import type * as Attestation from '../attestation/Types.js'
-import type { Constants } from './Constants.js'
-
 /** Verified Web Bot Auth request signature. */
 export type VerifiedRequest = {
   /** Public-key thumbprint advertised by the bot. */
@@ -10,6 +7,3 @@ export type VerifiedRequest = {
   /** Signature nonce supplied by the bot. */
   nonce: string
 }
-
-/** Web Bot Auth evidence made available to generic attestation policies. */
-export type Evidence = Attestation.Evidence<typeof Constants.protocol, VerifiedRequest>
