+++ test/tsconfig.json
@@ -14,11 +14,7 @@
       "mppx/server": ["../src/server/index.ts"],
       "mppx/tempo": ["../src/tempo/index.ts"],
       "mppx/attestation/tap": ["../src/tap/index.ts"],
-      "mppx/attestation/tap/client": ["../src/tap/client/index.ts"],
-      "mppx/attestation/tap/server": ["../src/tap/server/index.ts"],
       "mppx/attestation/web-bot-auth": ["../src/web-bot-auth/index.ts"],
-      "mppx/attestation/web-bot-auth/client": ["../src/web-bot-auth/client/index.ts"],
-      "mppx/attestation/web-bot-auth/server": ["../src/web-bot-auth/server/index.ts"],
       "mppx/x402": ["../src/x402/index.ts"],
       "mppx/hono": ["../src/middlewares/hono.ts"],
       "mppx/express": ["../src/middlewares/express.ts"],
