+++ vite.config.ts
@@ -5,17 +5,7 @@ import { playwright } from 'vp/test/browser-playwright'
 
 // Shared aliases used by both projects
 const alias = {
-  'mppx/attestation/tap/client': path.resolve(import.meta.dirname, 'src/tap/client'),
-  'mppx/attestation/tap/server': path.resolve(import.meta.dirname, 'src/tap/server'),
   'mppx/attestation/tap': path.resolve(import.meta.dirname, 'src/tap'),
-  'mppx/attestation/web-bot-auth/client': path.resolve(
-    import.meta.dirname,
-    'src/web-bot-auth/client',
-  ),
-  'mppx/attestation/web-bot-auth/server': path.resolve(
-    import.meta.dirname,
-    'src/web-bot-auth/server',
-  ),
   'mppx/attestation/web-bot-auth': path.resolve(import.meta.dirname, 'src/web-bot-auth'),
   'mppx/attestation': path.resolve(import.meta.dirname, 'src/attestation'),
   'mppx/client': path.resolve(import.meta.dirname, 'src/client'),
